requirements.txt
streamlit==1.32.2
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.2
PyPDF2==3.0.1
python-docx==1.1.2
